requirements.txt
fastapi
uvicorn
python-multipart
llama-index
torch
transformers
sentence-transformers
pinecone-client
